Move uri tests to right place
